conform some common standards
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@37083 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/httpd.c
@@ -35,8 +35,8 @@
 static int wfsendfile(int fd, off_t offset, size_t nbytes, webs_t wp);
 static char *wfgets(char *buf, int len, webs_t fp);
 static int wfprintf(webs_t fp, char *fmt, ...);
-static size_t wfwrite(void *buf, int size, int n, webs_t fp);
-static size_t wfread(void *buf, int size, int n, webs_t fp);
+static size_t wfwrite(void *buf, size_t size, size_t n, webs_t fp);
+static size_t wfread(void *buf, size_t size, size_t n, webs_t fp);
 static int wfclose(webs_t fp);
 static int wfflush(webs_t fp);
 #ifndef VALIDSOURCE
@@ -1959,7 +1959,7 @@ static size_t websWrite(webs_t wp, char *fmt, ...)
 	return ret;
 }
 
-static size_t wfwrite(void *buf, int size, int n, webs_t wp)
+static size_t wfwrite(void *buf, size_t size, size_t n, webs_t wp)
 {
 
 	FILE *fp = wp->fp;
@@ -1989,7 +1989,7 @@ static int wfsendfile(int fd, off_t offset, size_t nbytes, webs_t wp)
 
 }
 
-static size_t wfread(void *p, int size, int n, webs_t wp)
+static size_t wfread(void *p, size_t size, size_t n, webs_t wp)
 {
 	char *buf = (void *)p;
 	size_t ret;
@@ -2000,9 +2000,9 @@ static size_t wfread(void *p, int size, int n, webs_t wp)
 		ret = sslbufferread((struct sslbuffer *)fp, buf, n * size);
 #elif defined(HAVE_MATRIXSSL)
 		//do it in chains
-		int cnt = (size * n) / 0x4000;
-		int i;
-		int len = 0;
+		size_t cnt = (size * n) / 0x4000;
+		size_t i;
+		size_t len = 0;
 
 		for (i = 0; i < cnt; i++) {
 			len += matrixssl_read(fp, buf, 0x4000);
@@ -2012,7 +2012,7 @@ static size_t wfread(void *p, int size, int n, webs_t wp)
 
 		ret = len;
 #elif defined(HAVE_POLARSSL)
-		int len = n * size;
+		size_t len = n * size;
 		fprintf(stderr, "read ssl %d\n", len);
 		ret = ssl_read((ssl_context *) fp, (unsigned char *)buf, &len);
 #endif
